Refactor mod heirarchy to share scraper
